requirements.txt
asgiref==3.7.2
asttokens==2.4.1
Babel==2.14.0
decorator==5.1.1
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docxcompose==1.4.0
docxtpl==0.16.8
executing==2.0.1
gunicorn==21.2.0
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
load-dotenv==0.1.0
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
packaging==24.0
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
PyJWT==2.8.0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.10.0
wcwidth==0.2.13
whitenoise==6.6.0
